[CHG] refactoring global config
