+++ tests/files/config_s3imports.yml
@@ -1,0 +1,20 @@
+s3logsbeat:
+  inputs:
+    # S3 inputs (only taken into account when command `s3import` is executed)
+    -
+      type: s3
+      # S3
+      buckets:
+        - s3://mybucket/mypath
+      log_format: alb
+      # Optional fields extracter from key. E.g. key=staging-myapp/eu-west-1/2018/06/01/
+      key_regex_fields: ^(?P<environment>[^\-]+)-(?P<application>[^/]+)/(?P<awsregion>[^/]+)
+      since: 2018-10-15T01:00 # ISO8601 format - optional
+      to: 2018-11-20T01:00 # ISO8601 format - optional
+
+
+# Additional stuff we should be ignore
+output:
+  elasticsearch:
+    enabled: true
+    hosts: ["127.0.0.1:9200"]
